requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.2.3
WTForms==3.0.1
